refactor: deduplicate benchmark env parsing
Extract the repeated positive integer environment parsing logic into a shared bench utility.

Update the object diff, CRDT diff, and sequential apply microbenchmarks to import the shared helper instead of carrying local copies, and include a patch changeset describing the benchmark-only cleanup.
