+++ .changeset/late-gorillas-laugh.md
@@ -1,0 +1,5 @@
+---
+"json-patch-to-crdt": patch
+---
+
+Deduplicate benchmark environment parsing by introducing a shared `parsePositiveIntEnv` helper used across the object diff, CRDT diff, and sequential apply microbenchmarks.
