+++ bench/object-diff.microbench.ts
@@ -1,4 +1,5 @@
 import { diffJsonPatch, type JsonPatchOp, type JsonValue } from "../src/internals";
+import { parsePositiveIntEnv } from "./utils";
 
 type BenchmarkStats = {
   readonly name: string;
@@ -8,20 +9,6 @@ type BenchmarkStats = {
   readonly p50Ms: number;
   readonly avgHeapDeltaMb: number;
 };
-
-function parsePositiveIntEnv(name: string, fallback: number): number {
-  const raw = Bun.env[name];
-  if (!raw) {
-    return fallback;
-  }
-
-  const parsed = Number(raw);
-  if (!Number.isFinite(parsed) || parsed <= 0 || !Number.isInteger(parsed)) {
-    throw new Error(`${name} must be a positive integer, got '${raw}'`);
-  }
-
-  return parsed;
-}
 
 function percentile(sortedValues: number[], p: number): number {
   const pos = Math.floor((sortedValues.length - 1) * p);
