+++ bench/sequential-apply.microbench.ts
@@ -12,6 +12,7 @@ import {
   type JsonPatchOp,
   type JsonValue,
 } from "../src/internals";
+import { parsePositiveIntEnv } from "./utils";
 
 type BenchmarkStats = {
   name: string;
@@ -28,20 +29,6 @@ type BenchmarkScenario = {
   patchLength: number;
   runs: number;
 };
-
-function parsePositiveIntEnv(name: string, fallback: number): number {
-  const raw = Bun.env[name];
-  if (!raw) {
-    return fallback;
-  }
-
-  const parsed = Number(raw);
-  if (!Number.isFinite(parsed) || parsed <= 0 || !Number.isInteger(parsed)) {
-    throw new Error(`${name} must be a positive integer, got '${raw}'`);
-  }
-
-  return parsed;
-}
 
 function buildBaseDocument(size: number): JsonValue {
   return {
