+++ bench/utils.ts
@@ -1,0 +1,13 @@
+export function parsePositiveIntEnv(name: string, fallback: number): number {
+  const raw = Bun.env[name];
+  if (!raw) {
+    return fallback;
+  }
+
+  const parsed = Number(raw);
+  if (!Number.isFinite(parsed) || parsed <= 0 || !Number.isInteger(parsed)) {
+    throw new Error(`${name} must be a positive integer, got '${raw}'`);
+  }
+
+  return parsed;
+}
